Created a manager for searchHistory functionality to avoid controller complexity.
